Fix errors

+++ src/main/kotlin/com/gocode/webshop/usermanagement/service/UserService.kt
@@ -5,7 +5,6 @@ import com.gocode.webshop.constants.DELETED
 import com.gocode.webshop.usermanagement.model.Address
 import com.gocode.webshop.usermanagement.model.User
 import com.gocode.webshop.usermanagement.repository.UserRepository
-import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
 import java.lang.IllegalArgumentException
 import java.util.*
@@ -14,7 +13,7 @@ import java.util.*
 class UserService(
     private val userRepository: UserRepository,
     private val addressService: AddressService,
-    private val passwordEncoder: PasswordEncoder
+    //private val passwordEncoder: PasswordEncoder
 ) {
     fun findUserById(userId: UUID): User {
         return userRepository.findById(userId)
@@ -25,14 +24,16 @@ class UserService(
         return userRepository.save(
             user.copy(
                 id = null,
-                password = passwordEncoder.encode(user.password)
+                password = user.password
             )
         )
     }
 
     fun deleteUser(userId: UUID): User {
-        val user = userRepository.findById(userId)
-            .orElseThrow { throw EntityNotFoundException(userId.toString(), User::class.java) }
+        val user = findUserById(userId)
+        if (user.deleted) {
+            throw IllegalArgumentException("Can't delete already deleted user")
+        }
         deleteAddresses(userId)
         return userRepository.save(
             user.copy(
@@ -51,6 +52,9 @@ class UserService(
             throw IllegalArgumentException("Can't update user with null id")
         }
         val originalUser = findUserById(user.id)
+        if (originalUser.deleted) {
+            throw IllegalArgumentException("Can't update deleted user")
+        }
         return userRepository.save(
             originalUser.copy(
                 role = user.role,
@@ -64,32 +68,37 @@ class UserService(
 
     fun changePassword(userId: UUID, password: String): User {
         val originalUser = findUserById(userId)
+        if (originalUser.deleted) {
+            throw IllegalArgumentException("Can't change password of deleted user")
+        }
         return userRepository.save(
             originalUser.copy(
-                password = passwordEncoder.encode(password)
+                password = password
             )
         )
     }
 
     fun getAddresses(userId: UUID): List<Address> {
-        if (userRepository.existsById(userId)) {
-            return addressService.findByUserId(userId)
-        } else {
+        if (!userRepository.existsById(userId)) {
             throw EntityNotFoundException(userId.toString(), User::class.java)
         }
+        return addressService.findByUserId(userId)
     }
 
     fun deleteAddresses(userId: UUID) {
-        if (userRepository.existsById(userId)) {
-            val addresses = getAddresses(userId)
-            addresses.map { addressService.deleteAddress(it) }
-        } else {
+        if (!userRepository.existsById(userId)) {
             throw EntityNotFoundException(userId.toString(), User::class.java)
         }
+        val addresses = getAddresses(userId)
+        addresses.map { addressService.deleteAddress(it) }
     }
 
     fun findAllUsers(): List<User> {
         return userRepository.findAll()
+    }
+
+    fun findAllNonDeletedUsers(): List<User> {
+        return findAllUsers().filter { !it.deleted }
     }
 
 }
